Fix branch list including the upstream and any detached head.
If the upstream is a branch instead of a remote, it would be included
in the list of branches to rebase, likewise if a branch had already
been merged. Exclude those using --no-merged.
Use git for-each-ref instead of git branch to avoid any detached
head being added to the list.